requirements.txt
click==8.1.3
click-option-group==0.5.5
Cython==0.29.34
numpy==1.24.3
pysam==0.21.0
python-magic==0.4.27
